+++ predprof_teplici/app/static/js/toast.js
