+++ predprof_teplici/app/static/styles/ground.css
@@ -1,0 +1,51 @@
+.card{
+    border-radius: 10px;
+    border: 2px solid rgba(255, 255, 255, 0.5);
+    background: linear-gradient(to right bottom, rgba(255, 255, 255, 0.5), rgba(255, 255, 255, 0.75));
+    box-shadow: 0 2px 10px -5px black;
+    padding: 5px;
+}
+.card.ground{
+    width: 292;
+    height: 100%;
+    position: relative;
+    display: flex;
+    align-items: flex-end;
+}
+.card.ground .plant{
+    width: 115px;
+}
+.card.ground button{
+    width: 52px;
+    height: 52px;
+    margin: 0 5px;
+    box-shadow: 0 0 7px -5px black;
+    transition: 0.15s;
+    border-radius: 5px;
+}
+.card.ground  button:hover{
+    background-color: #d1cfcf;
+}
+.card.ground button img{
+    width: 100%;
+}
+.card.ground .data{
+    position: absolute;
+    top: 10px;
+    right: 0;
+    display: flex;
+    align-items: center;
+}
+.card.ground .data img{
+    height: 50px;
+}
+.card h2{
+    font-size: 2.5em;
+    min-width: 155px;
+}
+.card h3{
+    font-size: 1.3em;
+}
+.card h4{
+    font-size: 1.7em;
+}
