+++ predprof_teplici/app/static/styles/table.css
@@ -1,0 +1,38 @@
+
+        /*
+        <tr>
+            <th scope="col">Time</th>
+            <th scope="col">1</th>
+        </tr>
+        <tr>
+            <th scope="row">Monday</th>
+            <td>35</td>
+        </tr>
+        */
+td,
+th {
+    border: 1px solid rgb(136, 136, 136);
+    padding: 15px 25px;
+    font-size: 1.6em;
+}
+
+td {
+    text-align: center;
+}
+
+th[scope='col'] {
+    background-color: #559a78;
+    color: #fff;
+}
+
+th[scope='row'] {
+    background-color: #00ff003a;
+}
+
+table {
+    border-collapse: collapse;
+    border: 2px solid rgb(200, 200, 200);
+    letter-spacing: 1px;
+    font-family: sans-serif;
+    font-size: 0.8rem;
+}
