--- predprof_teplici/static/styles/form.css
@@ -1,119 +1,0 @@
-.form {
-    background-color: #0a3519;
-    border-radius: 20px;
-    box-sizing: border-box;
-    height: 500px;
-    padding: 20px;
-    width: 320px;
-    display: inline-block;
-    vertical-align: top;
-  }
-  
-  .title {
-    color: #eee;
-    font-family: sans-serif;
-    font-size: 36px;
-    font-weight: 600;
-    margin-top: 30px;
-  }
-  .subtitle {
-    color: #eee;
-    font-family: sans-serif;
-    font-size: 24px;
-    font-weight: 600;
-    margin-top: 30px;
-  }
-  
-  
-  .input-container {
-    height: 50px;
-    position: relative;
-    width: 100%;
-  }
-  
-  .form.big {
-      height: auto;
-      width: 400px;
-  }
-  .input-line{
-      display: flex;
-      flex-basis: auto;
-      width: 100%;
-      height: 50px;
-  }
-  .input-line > .input-container.small {
-    flex: auto;
-    flex-basis: auto;
-  }
-  .ic1 {
-    margin-top: 40px;
-  }
-  
-  .ic2 {
-    margin-top: 30px;
-  }
-  
-  .input {
-    background-color: #ffffff0e;
-    border-radius: 12px;
-    border: 0;
-    box-sizing: border-box;
-    color: #eee;
-    font-size: 18px;
-    height: 100%;
-    outline: 0;
-    padding: 4px 20px 0;
-    width: 100%;
-  }
-  
-  .placeholder {
-    color: rgb(108, 139, 113);
-    font-family: sans-serif;
-    left: 20px;
-    line-height: 14px;
-    pointer-events: none;
-    position: absolute;
-    transform-origin: 0 50%;
-    transition: transform 100ms, color 100ms;
-    top: 20px;
-  }
-  
-  .input:focus ~ .placeholder,
-  .input:not(:placeholder-shown) ~ .placeholder {
-    transform: translateY(-30px) translateX(10px) scale(0.75);
-  }
-  
-  .input:not(:placeholder-shown) ~ .placeholder {
-    color: rgb(108, 139, 113);
-  }
-  
-  .input:focus ~ .placeholder {
-    color: #06a3ae;
-  }
-  
-  .submit {
-    background-color: #08d;
-    border-radius: 12px;
-    border: 0;
-    box-sizing: border-box;
-    color: #eee;
-    cursor: pointer;
-    font-size: 18px;
-    height: 50px;
-    margin-top: 38px;
-    outline: 0;
-    text-align: center;
-    width: 100%;
-  }
-  
-  .submit:active {
-    background-color: #06b;
-    color: rgb(108, 139, 113);
-  }
-  .half{
-      width: 50%;
-  }
-  .data{
-      display: flex;
-  
-  }
