--- predprof_teplici/static/styles/modal.css
@@ -1,39 +1,0 @@
-
-  
-  .modal {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    position: absolute;
-    z-index: 2;
-    top: 5%;
-    background-color: rgba(0, 0, 0, 0);
-      
-    left: 50%;
-    transform: translate(-50%, 0);
-  }
-  
-  .modal .card {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 24px;
-  }
-  
-  .overlay {
-    position: fixed;
-    top: 0;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    width: 100%;
-    height: 100%;
-    background: rgba(0, 0, 0, 0.5);
-    backdrop-filter: blur(3px);
-    z-index: 1;
-  }
-  
-  .hidden {
-    display: none;
-  }
